backend/requirements.txt
amqp==5.0.6
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.1
celery-progress==0.1.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-mysql==3.8.1
django-timezone-field==4.2.1
djangorestframework==3.13.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
fastjsonschema==2.16.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kombu==5.2.2
lxml==4.6.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mysql-connector-python==8.0.26
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
nest-asyncio==1.5.6
notebook==6.4.12
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.9.2
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-environ==0.4.54
pytz==2021.3
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
redis==4.0.2
requests==2.26.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.25
sqlparse==0.4.2
stack-data==0.5.1
terminado==0.16.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.4.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.3
wrapt==1.13.3
zipp==3.8.1
